refactor(cockpit): move supply-chain gate into module
